feat: split into `@deeplx/core` and `@deeplx/cli`
close #17
