+++ .changeset/fresh-dingos-create.md
@@ -1,0 +1,7 @@
+---
+"@deeplx/core": minor
+"@deeplx/cli": minor
+"deeplx": minor
+---
+
+feat: split into `@deeplx/core` and `@deeplx/cli`
