+++ .codesandbox/ci.json
@@ -1,4 +1,7 @@
 {
   "node": "20",
+  "packages": [
+    "packages/**"
+  ],
   "sandboxes": []
 }
