+++ .github/workflows/pkg-pr-new.yml
@@ -28,4 +28,4 @@ jobs:
         run: yarn build
 
       - name: Publish
-        run: yarn dlx pkg-pr-new publish # --compact
+        run: yarn dlx pkg-pr-new publish './packages/**' # --compact
