+++ CHANGELOG.md
@@ -1,43 +1,8 @@
-# deeplx
+# Change Log
 
-## 0.2.1
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-### Patch Changes
-
-- [#29](https://github.com/un-ts/deeplx/pull/29) [`3760d7b`](https://github.com/un-ts/deeplx/commit/3760d7b1211aeb8d7cc00b68e23b462b6d4613ab) Thanks [@JounQin](https://github.com/JounQin)! - feat: change cli options and `formal` signature
-
-## 0.2.0
-
-### Minor Changes
-
-- [#25](https://github.com/un-ts/deeplx/pull/25) [`65cbb6d`](https://github.com/un-ts/deeplx/commit/65cbb6dd1191da883e7db039bbcce69d09dc2e7a) Thanks [@JounQin](https://github.com/JounQin)! - feat!: rewrite from scratch, port `OwO-Network/DeepLX` instead, migrate to `x-fetch`
-
-### Patch Changes
-
-- [#27](https://github.com/un-ts/deeplx/pull/27) [`6c371da`](https://github.com/un-ts/deeplx/commit/6c371da7cc8d2b4b538eb715d6ce414c2a005f2e) Thanks [@JounQin](https://github.com/JounQin)! - feat: add Ukrainian support
-
-## 0.1.3
-
-### Patch Changes
-
-- [#13](https://github.com/un-ts/deeplx/pull/13) [`d668994`](https://github.com/un-ts/deeplx/commit/d66899410976f1fb57816b5a1926b87b20615aff) Thanks [@niizam](https://github.com/niizam)! - Add 'Indonesian' support
-
-## 0.1.2
-
-### Patch Changes
-
-- [#7](https://github.com/un-ts/deeplx/pull/7) [`b24ab73`](https://github.com/un-ts/deeplx/commit/b24ab73c26c80db240a2d63936e00c96a010b507) Thanks [@toprak](https://github.com/toprak)! - feat: add Turkish support
-
-## 0.1.1
-
-### Patch Changes
-
-- [#4](https://github.com/un-ts/deeplx/pull/4) [`f7a6eb8`](https://github.com/un-ts/deeplx/commit/f7a6eb8c1cd6f069969db4667f195e0ae195a34b) Thanks [@JounQin](https://github.com/JounQin)! - feat: handle empty text earlier
-
-## 0.1.0
-
-### Minor Changes
-
-- [`b65e4d7`](https://github.com/un-ts/deeplx/commit/b65e4d77ad168dcb1962053bf8545f135074ffab) Thanks [@JounQin](https://github.com/JounQin)! - refactor: support `auto` for source language
-
-* [`0882e36`](https://github.com/un-ts/deeplx/commit/0882e364ffcc2efe2ec68cfe6e3284af82a47714) Thanks [@JounQin](https://github.com/JounQin)! - feat: first blood, should just work
+- [@deeplx/cli](./packages/@deeplx/cli/CHANGELOG.md)
+- [@deeplx/core](./packages/@deeplx/core/CHANGELOG.md)
+- [deeplx](./packages/deeplx/CHANGELOG.md)
