+++ eslint.config.js
@@ -1,8 +1,1 @@
-import base from '@1stg/eslint-config'
-
-export default [
-  ...base,
-  {
-    ignores: ['api/_deeplx.js'],
-  },
-]
+export { default } from '@1stg/eslint-config'
