+++ packages/@deeplx/cli/README.md
@@ -1,11 +1,11 @@
-# DeepL Translate
+# @deeplx/cli
 
 [![GitHub Actions Workflow Status](https://img.shields.io/github/actions/workflow/status/un-ts/deeplx/ci.yml?branch=master)](https://github.com/un-ts/deeplx/actions/workflows/ci.yml?query=branch%3Amaster)
 [![Codecov](https://img.shields.io/codecov/c/github/un-ts/deeplx.svg)](https://codecov.io/gh/un-ts/deeplx)
 [![Codacy Grade](https://img.shields.io/codacy/grade/e3d752491a664d889c5bdfb6ffeb6cbb)](https://app.codacy.com/gh/un-ts/deeplx)
 [![type-coverage](https://img.shields.io/badge/dynamic/json.svg?label=type-coverage&prefix=%E2%89%A5&suffix=%&query=$.typeCoverage.atLeast&uri=https%3A%2F%2Fraw.githubusercontent.com%2Frx-ts%2Fdeeplx%2Fmaster%2Fpackage.json)](https://github.com/plantain-00/type-coverage)
 [![CodeRabbit Pull Request Reviews](https://img.shields.io/coderabbit/prs/github/un-ts/deeplx)](https://coderabbit.ai)
-[![npm](https://img.shields.io/npm/v/deeplx.svg)](https://www.npmjs.com/package/deeplx)
+[![npm](https://img.shields.io/npm/v/@deeplx/cli.svg)](https://www.npmjs.com/package/@deeplx/cli)
 [![GitHub Release](https://img.shields.io/github/release/un-ts/deeplx)](https://github.com/un-ts/deeplx/releases)
 
 [![Conventional Commits](https://img.shields.io/badge/conventional%20commits-1.0.0-yellow.svg)](https://conventionalcommits.org)
@@ -14,7 +14,7 @@
 [![Code Style: Prettier](https://img.shields.io/badge/code_style-prettier-ff69b4.svg)](https://github.com/prettier/prettier)
 [![changesets](https://img.shields.io/badge/maintained%20with-changesets-176de3.svg)](https://github.com/atlassian/changesets)
 
-An unofficial Node package to translate text using [DeepL](https://www.deepl.com) by porting [OwO-Network/DeepLX](https://github.com/OwO-Network/DeepLX).
+The cli for [`@deeplx/core`](https://github.com/un-ts/deeplx/blob/master/packages/@deeplx/core).
 
 ## Online Service
 
@@ -24,15 +24,13 @@ An unofficial Node package to translate text using [DeepL](https://www.deepl.com
 
 ```sh
 # npm
-npm i deeplx
+npm i @deeplx/cli
 
 # pnpm
-pnpm add deeplx
+pnpm add @deeplx/cli
 
 # yarn
-yarn add deeplx
-
-# or install `@deeplx/core` for only API usage, or install `@deeplx/cli` for only CLI usage
+yarn add @deeplx/cli
 ```
 
 ## Usage
@@ -71,9 +69,7 @@ Currently the following languages are supported:
 
 You can either input the abbreviation or the language written in english.
 
-### Command line tool
-
-#### Help
+### Help
 
 ```sh
 deeplx -h
@@ -95,7 +91,7 @@ Options:
   -h, --help           display help for command
 ```
 
-#### Example 1
+### Example 1
 
 This will translate a Spanish (`ES`) text into Russian (`RU`):
 
@@ -107,7 +103,7 @@ deeplx -t russian --text "¡Buenos días!"
 Доброе утро!
 ```
 
-#### Example 2
+### Example 2
 
 This will translate the file (`test.txt`) text from Italian (`IT`) into Portuguese (`PT`):
 
@@ -115,7 +111,7 @@ This will translate the file (`test.txt`) text from Italian (`IT`) into Portugue
 deeplx -t PT -f test.txt
 ```
 
-#### Example 3
+### Example 3
 
 This will translate a Spanish (`ES`) text into Russian (`RU`) in _formal_ tone:
 
@@ -129,7 +125,7 @@ deeplx -t RU --text "¿Cómo te llamas?" --formal
 
 Note: _informal_ would be "_Как **тебя** зовут?_"
 
-#### Example 4
+### Example 4
 
 This will translate a Japanese (`JA`) text into German (`DE`) in _informal_ tone:
 
@@ -142,36 +138,6 @@ Wie geht es dir?
 ```
 
 Note: _formal_ would be "_Wie geht es **Ihnen**?_"
-
-### Node library
-
-#### Example 1
-
-This will translate a Chinese (`ZH`) text into Dutch (`NL`):
-
-```js
-import { translate } from 'deeplx'
-
-await translate('你好', 'NL')
-```
-
-```log
-'Hallo'
-```
-
-#### Example 2
-
-This will translate a `danish` text into `german` in informal tone:
-
-```js
-import { translate } from 'deeplx'
-
-await translate('Ring til mig!', 'german', 'danish', false)
-```
-
-```log
-'Ruf mich an!'
-```
 
 [![Sponsors](https://raw.githubusercontent.com/1stG/static/master/sponsors.svg)](https://github.com/sponsors/JounQin)
