+++ packages/@deeplx/core/index.d.cts
@@ -1,0 +1,3 @@
+import * as core from './lib/index.js'
+
+export = core
