+++ packages/deeplx/src/cli.ts
@@ -1,0 +1,1 @@
+export * from '@deeplx/cli'
