--- src/index.ts
@@ -1,5 +1,0 @@
-export * from './api.js'
-export * from './constants.js'
-export * from './translate.js'
-export type * from './types.js'
-export * from './utils.js'
