+++ test/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "@1stg/tsconfig/node16",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "include": ["../packages/**/test/**/*.spec.ts", "./global.d.ts"]
+}
