+++ vitest.config.ts
@@ -1,17 +1,14 @@
+import tsconfigPaths from 'vite-tsconfig-paths'
 import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
-  resolve: {
-    alias: {
-      deeplx: new URL('src/index.ts', import.meta.url).pathname,
-    },
-  },
+  plugins: [tsconfigPaths({})],
   test: {
     globals: true,
     coverage: {
       enabled: true,
-      include: ['src'],
-      exclude: ['src/cli.ts', 'src/fetch.ts'],
+      include: ['packages/**/src'],
+      exclude: ['packages/**/src/cli.ts', 'packages/**/src/fetch.ts'],
       provider: 'istanbul',
       reporter: ['lcov', 'json', 'text'],
     },
